mueve la sección GraphQL a una nueva página
