+++ react/src/pages/graphql/index.tsx
@@ -1,0 +1,37 @@
+import { useQueryObtenerEmpresas } from "./querys";
+
+const CargandoDatos = ({ texto }: { texto?: string }) => <p>Cargando {texto}</p>;
+const ErrorComponente = () => <p>Ocurrió un error</p>;
+
+export const GraphqlPage = () => {
+  
+  const [result] = useQueryObtenerEmpresas();
+  const { data, error, fetching } = result;
+
+  return (
+    <>
+
+      {
+        fetching && (
+          <CargandoDatos texto="compañias" />
+        )
+      }
+
+      {
+        error && (
+          <ErrorComponente />
+        )
+      }
+
+      {
+        !fetching && data.allCompanies.length > 0 && data.allCompanies.map((item: any) => {
+          return (
+            <li>{item.name}</li>
+          )
+        })
+      }
+    </>
+  );
+};
+
+export default GraphqlPage;
